Add files via upload
Classification of gene expression data has been a booming topic in the recent years. This can help
develop efficient methods in the field of bioinformatics to be used for cancer diagnosis and further, it’s
treatment. This seminar gives a critical review of existing data mining techniques being practiced in the
field of analysis of gene expression data. Due to the complexity of the underlying biological processes,
the mining of gene expression data for the purpose of gene function prediction becomes very difficult.
A fuzzy mining technique is explained to overcome these difficulties. This can effectively capture dis-
parateness in expression data for discovery of patterns by transforming quantitative expression values
into linguistic terms, such as highly or lowly expressed. It makes use of a fuzzy measure to determine
if interesting association patterns exist between the linguistic gene expression levels. It can be used to
expose hidden patterns to accurately predict and classify gene functions.